+++ stroom-expression/src/main/java/stroom/dashboard/expression/ExtractionFunction.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2018 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package stroom.dashboard.expression;
+
+import java.io.Serializable;
+import java.text.ParseException;
+
+public abstract class ExtractionFunction extends AbstractFunction implements Serializable {
+    private Generator gen;
+    private Function function = null;
+    private boolean hasAggregate;
+
+    public ExtractionFunction(final String name) {
+        super(name, 1, 1);
+    }
+
+    @Override
+    public void setParams(final Object[] params) throws ParseException {
+        super.setParams(params);
+
+        final Object param = params[0];
+        if (param instanceof Function) {
+            function = (Function) param;
+            hasAggregate = function.hasAggregate();
+        } else {
+            // Optimise replacement of static input in case user does something stupid.
+            final String string = param.toString();
+            gen = new StaticValueFunction(getExtractor().extract(string)).createGenerator();
+            hasAggregate = false;
+        }
+    }
+
+    @Override
+    public Generator createGenerator() {
+        if (gen != null) {
+            return gen;
+        }
+
+        final Generator childGenerator = function.createGenerator();
+        return new ExtractionGenerator(childGenerator, getExtractor());
+    }
+
+    @Override
+    public boolean hasAggregate() {
+        return hasAggregate;
+    }
+
+    abstract Extractor getExtractor();
+}