+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#554** : Added a `parseUri` XSLT function.
 
+* Issue **#557** : Added dashboard functions to parse and output URI parts.
+
 ## [v5.1-beta.8] - 2018-01-12
 
 * Issue **#550** : Fix for info message output for data retention.